backend/requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
boto3==1.42.66
botocore==1.42.66
celery==5.6.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==6.0.3
django-cors-headers==4.9.0
django-storages==1.14.6
djangorestframework==3.16.1
jmespath==1.1.0
kombu==5.6.2
packaging==26.0
pillow==12.1.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
redis==6.4.0
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
